requirements.txt
# Whisper for transcription
git+https://github.com/openai/whisper.git

# YouTube downloader
yt-dlp

# Sentence Transformers for embeddings
sentence-transformers

# PyTorch (CPU version, adjust if you have GPU)
torch
pydub
groq
python-dotenv